+++ src_myfolder/features/posts/ui/components/PostManagerHeader.tsx
