--- src_myfolder/widgets/index.ts
@@ -1,5 +1,0 @@
-// 복합 위젯들만 export (여러 features를 조합하거나 완성된 UI 블록)
-export { PostsTable } from "./posts-table/ui/PostsTable"
-export { default as PaginationControls } from "./pagination/ui/PaginationControls"
-export { default as PostFilter } from "./post-filter/ui/PostFilter"
-export { default as PostManagerHeader } from "./post-manager-header/ui/PostManagerHeader"
